templatized inter data structures to allow double and long double as the underlying types

+++ args.hpp
@@ -8,7 +8,7 @@ struct Args {
   static unsigned d ; // -d dims: an integer specifying the dimension of the points
   static std::string  i; // -i inputfilename: a string specifying the input filename
   static unsigned m; // -m max_num_iter: an integer specifying the maximum number of iterations
-  static double t; // -t threshold: a double specifying the threshold for convergence test.
+  static long double t; // -t threshold: a double specifying the threshold for convergence test.
   static bool c; // -c: a flag to control the output of your program. 
                 // If -c is specified, your program should output the centroids of all clusters. 
                 // If -c is not specified, your program should output the labels of all points.
